+++ src/ts/components/settings/SettingsPanel.ts
@@ -42,6 +42,12 @@ export interface SettingsPanelConfig extends ContainerConfig {
    * Default: true
    */
   hideOnControlsHide?: boolean;
+
+  /**
+   * Specifies if the settings panel should be hidden when another settings panel is opened.
+   * Default: true
+   */
+  hideOnOtherSettingsPanelOpening?: boolean;
 }
 
 /**
@@ -114,6 +120,7 @@ export class SettingsPanel<Config extends SettingsPanelConfig> extends Container
         pageTransitionAnimation: true,
         stateResetDelay: 5000,
         hideOnControlsHide: true,
+        hideOnOtherSettingsPanelOpening: true,
       } as Config,
       this.config,
     );