+++ src/ts/localization/i18n.ts
@@ -50,6 +50,8 @@ export interface Vocabulary {
   'settings.audio.volume': string;
   'settings.subtitles': string;
   'settings.subtitles.options': string;
+  'playerInsights.show': string;
+  'playerInsights.hide': string;
   'settings.subtitles.font.color': string;
   'settings.subtitles.font.opacity': string;
   'settings.subtitles.background.color': string;