+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - New `ContextMenu` component to show custom context menus on right-click.
 - `UIInstanceManager.onActive` and `UIInstanceManager.onInactive` lifecycle events for UI variant status changes.
 - Default `PlayerContextMenu` to access player information directly from the UI.
+- `PlayerInsightsPanel` to show detailed player insights and diagnostics.
 
 ## [4.15.1] - 2026-06-11